# backend/api-documentation/openapi.yaml
openapi: 3.0.4
info:
  title: Swagger Gradia - OpenAPI 3.0
  description: >-
    This is a work in progress documentation for gradia REST API
  contact:
    email: ""

  version: 1.0.12
servers:
  - url: https://backend-511443014815.europe-west1.run.app
...
paths:
...
  /tag:
    post:
      security:
        - cookieAuth: [ ]
      tags:
        - tag
      summary: Create tag
      description: Creates a new tag
      operationId: createTag
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTag'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      security:
        - cookieAuth: [ ]
      tags:
        - tag
      summary: Get all existing tags
      operationId: getAllTags
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagArray'
        '400':
          description: Invalid id supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tag/{tagId}:
    get:
      security:
        - cookieAuth: [ ]
      tags:
        - tag
      summary: Get tag by id.
      description: Get tag detail based on id.
      operationId: getTagById
      parameters:
        - name: tagId
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          description: Invalid id supplied
        '404':
          description: User not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      security:
        - cookieAuth: [ ]
      tags:
        - tag
      summary: Update tag resource.
      description: This can only be done by the logged in user.
      operationId: updateTag
      parameters:
        - name: tagId
          in: path
          description: id that need to be modified
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent user in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'

      responses:
        '200':
          description: successful operation
        '400':
          description: bad request
        '404':
          description: tag not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Tag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          enum:
            - blue
            - purple
            - yellow
            - green
            - red
            - pink
          example: blue
    CreateTag:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
          enum:
            - blue
            - purple
            - yellow
            - green
            - red
            - pink
          example: blue
    TagArray:
      type: array
      items:
        $ref: '#/components/schemas/Tag'
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid